Add a way to delay interpreting the SKX structure until actually needed.
In some case, with efficient servers, we end up parsing the SKX structure
without having yet set the pending cipher. The cipher key exchange type
being unknown at this stage, lead to not knowing how to parse the SKX
structure.

Fix it by keeping the byte un-interpreted when the cipher key exchange
is not known, and properly parse later on.

Fix #53.