Allow for flexible compilation against different network stacks.
This patch should be completely backwards compatible with existing systems
by default. However, it provides the ability (via Cabal flags) for the TLS
library to be compiled against an alternate network stack (HaNS), or
against both HaNS and the network library. It may be worth adding
additional backends later, as they arise.