Fix bug related to RSA signature of CertificateVerify with TLS < 1.2
This is mirror to what was already fixed in commit 638680ea.

For SSLv3, TLS 1.0, TLS 1.1 the SHA1_MD5 hashing was applied twice when
the client sends its certificate.

In certificateVerifyCreate we need to call a new function
signatureCreateWithHashDescr without this additional hashing.  Function
signatureCreate should not change because it is also called by the (EC)DH
signing functions.

The new function name is similar to what already exists for signature
verification, and design is now more symmetrical.