Take into account HashAndSignatureAlgorithms when selecting ciphers
Now that signature_algorithms extension is used to sign the
ServerKeyXchg, this creates a possibility that client and
server don't agree on the hash algorithm.  When this happens
a cipher with a different key-exchange should have been
selected.

We now skip ciphers when the key-exchange is not possible
because of this reason.