Strip leading zeros in DHE premaster secret
Restores the D-H key as it was used before commit e77b267, i.e. without
leading 0x00 bytes.  This is required by RFC 4346 (section 8.1.2) and applies
only to DH(E) up to TLS 1.2, specifically *not* ECDH(E).

Allows to fix an infrequent and random BadRecordMac failure during handshake,
when DHE is selected and the premaster secret began with a 0x00 byte.

Reported by Eric Hohenstein in #197.