Add missing checks when resuming a session
In RFC 5246 section 7.4.1.2 explains cipher_suites must contain the
cipher_suite from the resumed session.  Proceeds with a full handshake
otherwise, so a different cipher might be selected.

Althought not required, this commit also checks sessionVersion.
This looks safer, as there is a relationship between TLS version
and available ciphers.