Remove functions ctxCiphers and getCiphers
Embedding the logic to filter server ciphers directly in the parameters
themselves seems convoluted.  Moreover it applies only to server ciphers
and client ciphers do not need the extraCreds.

As credentials need to be filtered based on ClientHello extensions
it's best to move the function getCiphers directly in server code.