Check only leaf certificate from certificate chain
Draft TLS 1.3 section 4.4.2.2:  Certificates that are self-signed or
certificates that are expected to be trust anchors are not validated
as part of the chain and therefore MAY be signed with any algorithm.

This implementation is positional only.  Server does not necessarily
knows what level in the chain the client considers as trust anchor,
so the code evaluates only a leaf certificate that is not self-signed.

Test suite does not use self-signed certificates anymore.