Test resumption with separate IORefs for client/server
With only one IORef in test cases, the client was able to read the
IORef set by the server context, even when the client did not receive
the session data from the server.  So with this design we could not
detect properly all resumption issues.  Additionally, with TLS13,
client and server do not share exact same content in SessionData so
having distinct IORef storage is required.

This commit just puts two IORefs where we had only one, with some
appropriate utility functions.