Some thread-safety for writes
Calls to sendPacket/sendPacket13 are protected by the write lock to
make sure we never try to send two packets simultaneously.

In recvData, the use of the read lock is extended to the actual
processing of the packet just received.  This prevents another packet
to slip in before the full response is completed.