Use bidirectional "close_notify" in test suite
Application code which does not call 'recvData' just before 'bye' is
modified to wait for the peer "close_notify" message.  This ensures
that a test failure on one side of the connection is detected by the
other side even when the latter only sends data.  With this, session
tickets are also received correctly by a send-only client.