Don't send 0-RTT data when ticket is expired
The condition with isAgeValid was used for extension "pre_shared_key"
but not when determining whether to send 0-RTT data or not.  With an
expired ticket adjustExtentions was called and modified an extension
which was not "pre_shared_key".

This commit rearranges the variables so that the result of isAgeValid
is used at all necessary places.