Use RSA key length when considering hash algorithm
Now that RSASSA-PSS is enabled, and used by default with SHA-512, we
require an RSA key with 1034 bits minimum.  This may create failures
for people still using a 1024-bit key.

This commit adds the key-length criteria when selecting the hash
algorithm so that the selection is compatible with the actual key
length available.  It deals with both RSASSA-PKCS1 and RSASSA-PSS.
The (EC)DSA signature algorithms internally truncate the digest and
need no specific condition.